[ffmpeg_consumer] Retired old implementation in favour of the now updated streaming_consumer
