+++ shell/casparcg.config
@@ -107,17 +107,12 @@
                 <borderless>false [true|false]</borderless>
             </screen>
             <newtek-ivga></newtek-ivga>
-            <file>
-                <path></path>
-                <vcodec>libx264 [libx264|qtrle]</vcodec>
-                <separate-key>false [true|false]</separate-key>
-            </file>
-            <stream>
-                <path>udp://localhost:9250</path>
-                <args>-format mpegts -vcodec libx264 -crf 25 -tune zerolatency -preset ultrafast</args>
+            <ffmpeg>
+                <path>[file|url]</path>
+                <args>[most ffmpeg arguments related to filtering and output codecs]</args>
                 <separate-key>false [true|false]</separate-key>
                 <mono-streams>false [true|false]</mono-streams>
-            </stream>
+            </ffmpeg>
             <syncto>
                 <channel-id>1</channel-id>
             </syncto>
